refactor workspace helpers into utils and constants modules
